refactor: Split PluginTests.cs into focused test files
Split the monolithic PluginTests.cs (1,249 lines, 12+ test classes) into
smaller, focused test files for improved maintainability and navigation.

New files:
- PluginTestHelpers.cs - Shared test helper classes and plugins
- PluginLifecycleTests.cs - Installation, uninstallation, and dispose tests
- PluginSystemIntegrationTests.cs - System registration and integration tests
- ExtensionTests.cs - World extension API and plugin extension tests
- PluginQueryTests.cs - Plugin query method tests
- WorldBuilderTests.cs - WorldBuilder and dependency constraint tests
- PluginIsolationTests.cs - World isolation and context dependency tests

The split improves:
- Navigation and discoverability of specific tests
- Reduced merge conflict likelihood
- Faster IDE loading of test files
- Single responsibility per file

All 10,003 existing tests continue to pass.

Closes #473
